refactor(config): move global config accessors
