+++ crates/ecstore/src/config/mod.rs
@@ -34,14 +34,16 @@ use rustfs_config::notify::{
     NOTIFY_POSTGRES_SUB_SYS, NOTIFY_PULSAR_SUB_SYS, NOTIFY_REDIS_SUB_SYS, NOTIFY_WEBHOOK_SUB_SYS,
 };
 use rustfs_config::oidc::IDENTITY_OPENID_SUB_SYS;
-use rustfs_config::server_config::{Config, register_default_kvs};
+use rustfs_config::server_config::register_default_kvs;
 use std::collections::HashMap;
 use std::sync::LazyLock;
 use std::sync::{Arc, RwLock};
 
+// RUSTFS_COMPAT_TODO(CFG-008): keep old ecstore global server-config accessor path while runtime consumers migrate. Remove after all consumers import these accessors from rustfs_config::server_config.
+pub use rustfs_config::server_config::{get_global_server_config, set_global_server_config};
+
 pub static GLOBAL_STORAGE_CLASS: LazyLock<RwLock<storageclass::Config>> =
     LazyLock::new(|| RwLock::new(storageclass::Config::default()));
-pub static GLOBAL_SERVER_CONFIG: LazyLock<RwLock<Option<Config>>> = LazyLock::new(|| RwLock::new(None));
 pub static GLOBAL_CONFIG_SYS: LazyLock<ConfigSys> = LazyLock::new(ConfigSys::new);
 
 pub static RUSTFS_CONFIG_PREFIX: &str = "config";
@@ -66,16 +68,6 @@ impl ConfigSys {
         set_global_server_config(cfg);
 
         Ok(())
-    }
-}
-
-pub fn get_global_server_config() -> Option<Config> {
-    GLOBAL_SERVER_CONFIG.read().ok().and_then(|guard| (*guard).clone())
-}
-
-pub fn set_global_server_config(cfg: Config) {
-    if let Ok(mut guard) = GLOBAL_SERVER_CONFIG.write() {
-        *guard = Some(cfg);
     }
 }
 
@@ -132,7 +124,7 @@ pub fn init() {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use rustfs_config::server_config::KVS;
+    use rustfs_config::server_config::{Config, KVS};
     use rustfs_config::{
         DEFAULT_DELIMITER, DEFAULT_HEAL_BITROT_CYCLE_SECS, DEFAULT_SCANNER_SPEED, HEAL_BITROT_CYCLE, SCANNER_CYCLE_MAX_OBJECTS,
         SCANNER_DELAY, SCANNER_MAX_WAIT, SCANNER_SPEED, SCANNER_SUB_SYS,
